refactor(web): remove event utils barrel
