+++ apps/web/src/features/guild/events/components/dialogs/event-participation-confirmation-dialog.tsx
@@ -20,7 +20,7 @@ import {
   useConfirmParticipationForKill,
   useListPendingParticipationConfirmations,
 } from "@/lib/api/generated/main/events/events";
-import { formatDateTime } from "../../utils";
+import { formatDateTime } from "../../utils/format-date";
 import { invalidateKillQueries } from "../../hooks/mutations/invalidate-kill-queries";
 
 interface EventParticipationConfirmationDialogProps {
